requirements.txt
# ============================================
# APEX AutoVision - Complete Requirements
# ============================================

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers & NLP
transformers>=4.30.0
tokenizers>=0.13.0
sentencepiece>=0.1.99

# Computer Vision
timm>=0.9.0
pillow>=10.0.0
opencv-python>=4.8.0

# Machine Learning & Data Science
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0

# Statistical Analysis & Drift Detection
scipy>=1.10.0
statsmodels>=0.14.0

# API & Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Frontend & Visualization
streamlit>=1.28.0
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Data Storage & Serialization
joblib>=1.3.0
h5py>=3.9.0
pyarrow>=13.0.0

# HTTP Requests
requests>=2.31.0
aiohttp>=3.8.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Testing (Optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code Quality (Optional)
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Jupyter (Optional - for experimentation)
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.1.0